requirements.txt
astroid==3.3.10
dill==0.4.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
MouseInfo==0.1.3
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.2
pylint==3.3.7
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.1
python3-xlib==0.15
pytweening==1.2.0
tomlkit==0.13.3
